+++ packages/vseed/src/types/properties/encoding/zMeasureEncoding.ts
@@ -1,0 +1,16 @@
+import { z } from 'zod'
+
+export const zMeasureEncoding = z.enum([
+  'primaryYAxis',
+  'secondaryYAxis',
+  'xAxis',
+  'yAxis',
+  'angle',
+  'radius',
+  'size',
+  'color',
+  'detail',
+  'column',
+  'label',
+  'tooltip',
+])