+++ packages/vseed/src/types/properties/measures/measures.ts
@@ -1,3 +1,4 @@
+import type { MeasureEncoding } from '../encoding'
 import type { NumFormat } from './format/numFormat'
 
 /**
@@ -43,19 +44,7 @@ export type Measure = {
    * - label: 指标映射的标签, 适用于所有图表
    * - tooltip: 指标映射的提示, 适用于所有图表
    */
-  encoding?:
-    | 'primaryYAxis'
-    | 'secondaryYAxis'
-    | 'xAxis'
-    | 'yAxis'
-    | 'angle'
-    | 'radius'
-    | 'size'
-    | 'color'
-    | 'detail'
-    | 'column'
-    | 'label'
-    | 'tooltip'
+  encoding?: MeasureEncoding
 
   /**
    * @description 以扁平的指标配置形式, 构建树形指标组, parentId指向父级指标组的id, 用于构建指标树